feat: move /try-on route into separate file and use new image processing method
